values moved from constants to types folder
