+++ src/common/constants.tsx
@@ -909,78 +909,3 @@ export const HEADER_CONTENT_TYPES = {
 } as const;
 
 export const ADMIN_USER_TYPES = ["DistrictAdmin", "StateAdmin"] as const;
-
-export const ENCOUNTER_DIET_PREFERENCE = [
-  "vegetarian",
-  "diary_free",
-  "nut_free",
-  "gluten_free",
-  "vegan",
-  "halal",
-  "kosher",
-  "none",
-] as const;
-
-export const ENCOUNTER_STATUS = [
-  "planned",
-  "in_progress",
-  "on_hold",
-  "discharged",
-  "completed",
-  "cancelled",
-  "discontinued",
-  "entered_in_error",
-  "unknown",
-] as const;
-
-export const ENCOUNTER_CLASS = [
-  "imp",
-  "amb",
-  "obsenc",
-  "emer",
-  "vr",
-  "hh",
-] as const;
-
-export const ENCOUNTER_PRIORITY = [
-  "ASAP",
-  "callback_results",
-  "callback_for_scheduling",
-  "elective",
-  "emergency",
-  "preop",
-  "as_needed",
-  "routine",
-  "rush_reporting",
-  "stat",
-  "timing_critical",
-  "use_as_directed",
-  "urgent",
-] as const;
-
-export const ENCOUNTER_ADMIT_SOURCE = [
-  "hosp_trans",
-  "emd",
-  "outp",
-  "born",
-  "gp",
-  "mp",
-  "nursing",
-  "psych",
-  "rehab",
-  "other",
-] as const;
-
-export const ENCOUNTER_DISCHARGE_DISPOSITION = [
-  "home",
-  "alt_home",
-  "other_hcf",
-  "hosp",
-  "long",
-  "aadvice",
-  "exp",
-  "psy",
-  "rehab",
-  "snf",
-  "oth",
-] as const;
